aidl: fingerprint: Fix type mismatch in FingerprintCallback::onFingerprintCmd method
* This commit resolves a compilation error caused by a type mismatch between the declaration and implementation of the `onFingerprintCmd` method in the `FingerprintCallback` class.
* The second parameter was incorrectly defined as `const hidl_vec<int8_t>&` in both the implementation and the header file. It has been updated to `const hidl_vec<uint32_t>&` to match the expected type.

* The following error was encountered:

hardware/oplus/aidl/fingerprint/include/Session.h:293:55: error: non-virtual member function marked 'override' hides virtual member function
  293 |                                   uint32_t resultLen) override;
      |                                                       ^
out/soong/.intermediates/hardware/oplus/interfaces/oplus/hardware/biometrics/fingerprint/2.1/vendor.oplus.hardware.biometrics.fingerprint@2.1_genc++_headers/gen/vendor/oplus/hardware/biometrics/fingerprint/2.1/IBiometricsFingerprintClientCallbackEx.h:39:47:
note: hidden overloaded virtual function 'vendor::oplus::hardware::biometrics::fingerprint::V2_1::IBiometricsFingerprintClientCallbackEx::onFingerprintCmd'
declared here: type mismatch at 2nd parameter ('const ::android::hardware::hidl_vec<uint32_t> &' (aka 'const hidl_vec<unsigned int> &') vs 'const hidl_vec<int8_t> &' (aka 'const hidl_vec<signed char> &'))
   39 |     virtual ::android::hardware::Return<void> onFingerprintCmd(int32_t cmdId, const ::android::hardware::hidl_vec<uint32_t>& result, uint32_t resultLen) = 0;
      |                                               ^
1 error generated.

+++ aidl/fingerprint/Session.cpp
@@ -548,7 +548,7 @@ Return<void> FingerprintCallback::onEngineeringInfoUpdated(
 }
 
 Return<void> FingerprintCallback::onFingerprintCmd(int32_t cmdId,
-                                                     const hidl_vec<int8_t>& result,
+                                                     const hidl_vec<uint32_t>& result,
                                                      uint32_t resultLen) {
     switch (cmdId) {
         case FINGERPRINT_CALLBACK_CMD_ID_ON_TOUCH_DOWN:
